chore: cleaning file structure for primitive components
