--- packages/react/src/components/Box/Box.types.ts
@@ -1,8 +1,0 @@
-import type { CSS } from '@project44-manifest/react-styles';
-
-export type BoxElement = 'div';
-
-export interface BoxProps {
-  /** Theme aware style object */
-  css?: CSS;
-}
