--- packages/react/src/components/Container/index.ts
@@ -1,2 +1,0 @@
-export * from './Container';
-export * from './Container.types';
