--- packages/react/src/components/Flex/index.ts
@@ -1,2 +1,0 @@
-export * from './Flex';
-export * from './Flex.types';
