--- packages/react/src/components/Grid/index.ts
@@ -1,2 +1,0 @@
-export * from './Grid';
-export * from './Grid.types';
