--- packages/react/src/components/GridItem/GridItem.styles.ts
@@ -1,3 +1,0 @@
-import { styled } from '@project44-manifest/react-styles';
-
-export const StyledGridItem = styled('div', { boxSizing: 'border-box' });
