--- packages/react/src/components/GridItem/index.ts
@@ -1,2 +1,0 @@
-export * from './GridItem';
-export * from './GridItem.types';
