--- packages/react/src/components/Stack/Stack.tsx
@@ -1,21 +1,0 @@
-import * as React from 'react';
-import { cx } from '@project44-manifest/react-styles';
-import type { ForwardRefComponent } from '@project44-manifest/react-types';
-import { StyledStack } from './Stack.styles';
-import type { StackElement, StackProps } from './Stack.types';
-
-export const Stack = React.forwardRef((props, forwardedRef) => {
-  const { as, className: classNameProp, css, gap, orientation, ...other } = props;
-
-  return (
-    <StyledStack
-      {...other}
-      ref={forwardedRef}
-      as={as}
-      className={cx('manifest-stack', classNameProp)}
-      css={css}
-      gap={gap}
-      orientation={orientation}
-    />
-  );
-}) as ForwardRefComponent<StackElement, StackProps>;
