--- packages/react/src/components/Stack/index.ts
@@ -1,2 +1,0 @@
-export * from './Stack';
-export * from './Stack.types';
