+++ packages/react/src/components/container/index.ts
@@ -1,0 +1,1 @@
+export * from './container';
