+++ packages/react/src/components/flex/index.ts
@@ -1,0 +1,1 @@
+export * from './flex';
