+++ packages/react/src/components/grid/index.ts
@@ -1,0 +1,1 @@
+export * from './grid';
