+++ packages/react/src/index.ts
@@ -1,7 +1,6 @@
 export * from './components/action-card';
 export * from './components/avatar';
-export type { BoxProps } from './components/Box';
-export { Box } from './components/Box';
+export * from './components/box';
 export type { BreadcrumbProps } from './components/Breadcrumb';
 export { Breadcrumb } from './components/Breadcrumb';
 export type { BreadcrumbItemProps } from './components/BreadcrumbItem';
@@ -32,8 +31,7 @@ export type { ComboboxItemProps } from './components/ComboboxItem';
 export { ComboboxItem } from './components/ComboboxItem';
 export type { ComboboxSectionProps } from './components/ComboboxSection';
 export { ComboboxSection } from './components/ComboboxSection';
-export type { ContainerProps } from './components/Container';
-export { Container } from './components/Container';
+export * from './components/container';
 export type { DatePickerProps } from './components/DatePicker';
 export { DatePicker } from './components/DatePicker';
 export type { DateRangePickerProps } from './components/DateRangePicker';
@@ -56,14 +54,10 @@ export type { DropdownSectionProps } from './components/DropdownSection';
 export { DropdownSection } from './components/DropdownSection';
 export type { FadeProps } from './components/Fade';
 export { Fade } from './components/Fade';
-export type { FlexElement, FlexProps } from './components/Flex';
-export { Flex } from './components/Flex';
+export * from './components/flex';
 export type { FormControlProps } from './components/FormControl';
 export { FormControl } from './components/FormControl';
-export type { GridElement, GridProps } from './components/Grid';
-export { Grid } from './components/Grid';
-export type { GridItemElement, GridItemProps } from './components/GridItem';
-export { GridItem } from './components/GridItem';
+export * from './components/grid';
 export type { IconProps } from './components/Icon';
 export { Icon } from './components/Icon';
 export type { IconButtonElement, IconButtonProps } from './components/IconButton';
@@ -135,8 +129,7 @@ export type { SkeletonProps } from './components/Skeleton';
 export { Skeleton } from './components/Skeleton';
 export type { SpinnerProps } from './components/Spinner';
 export { Spinner } from './components/Spinner';
-export type { StackProps } from './components/Stack';
-export { Stack } from './components/Stack';
+export * from './components/stack';
 export type { SwitchProps } from './components/Switch';
 export { Switch } from './components/Switch';
 export type { TableProps } from './components/Table';
