+++ packages/react/tests/grid.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { Grid } from '../src';
 
-describe('react-layout - Grid', () => {
+describe('grid', () => {
   it('should render correctly', () => {
     render(<Grid>Hello World</Grid>);
